Dev issue 43 (#66)
Adding abort conditions logic.

Abort conditions will stop comparison for the field they are set up in case the condition fails, but will not mark the field as failed in comparison. This is helpful for comparison of fields that we can't have an expected for, like generated ID's. at the same time comparing them if we already know the value for a specific object. The first abort condition added is IgnoreComparisonIfExpectedNull